server/requirements.txt
aiosmtplib==3.0.2
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncmy==0.2.10
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-csrf-protect==1.0.7
fastapi-limiter==0.1.6
fastapi-mail==1.5.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pwdlib==0.2.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.3.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
